+++ src/main/java/org/example/springmvccrud/dao/UserDao.java
@@ -10,9 +10,9 @@ public interface UserDao {
 
     void updateUser(User user);
 
-    User getUserById(int id);
+    User getUserById(Long id);
 
-    void deleteUser(int id);
+    void deleteUser(Long id);
 
     List<User> getAllUsers();
 }
