+++ src/main/java/org/example/springmvccrud/dao/UserDaoImpl.java
@@ -26,12 +26,12 @@ public class UserDaoImpl implements UserDao {
     }
 
     @Override
-    public User getUserById(int id) {
+    public User getUserById(Long id) {
         return entityManager.find(User.class, id);
     }
 
     @Override
-    public void deleteUser(int id) {
+    public void deleteUser(Long id) {
         User user = getUserById(id);
         if (user != null) {
             entityManager.remove(user);
